requirements.txt
flask
torch
torchvision
open-clip-torch
pandas
pillow

